+++ Runtime/Adapters/JsonAdapter.SerializePrefab.cs
@@ -88,6 +88,7 @@ namespace Unity.RuntimeSceneSerialization.Json.Adapters
                 }
             }
 
+            // TODO: ARCC-116 This will skip hidden objects on Unity 6
             foreach (Transform child in transform)
             {
                 var childTransformPath = child.name;